+++ app/Http/Resources/Invoice/InvoiceResource.php
@@ -46,6 +46,9 @@ class InvoiceResource extends JsonResource
             'verification_url' => $this->when($this->uuid, function () {
                 return url('/api/v1/general/invoices/verify/' . $this->uuid);
             }),
+            'view_url' => $this->when($this->order_id, function () {
+                return url('/api/v1/general/orders/' . $this->order_id . '/invoice');
+            }),
         ];
     }